+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 1.15 (not yet released)
 
 * Added percentage calculation field to calculate the difference in percentage between two fields.
+* Fixed issue with saving default filter values.
 
 # Version 1.14.2