+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 1.15 (not yet released)
+# Version 1.15
 
 * Added percentage calculation field to calculate the difference in percentage between two fields.
 * Fixed issue with saving default filter values.